Release SoundMaster v0.9.3
Remove a maintainer's address from the build, repair video import, and stop
generation from failing on an engine the build cannot load.

A personal e-mail was hardcoded as the default publisher contact and therefore
shipped in every build. It is removed, and because it is already written into
legal_profile.json on existing installs it is also scrubbed from disk on load.
The compliance page becomes read-only: values are shown and can be copied, but
nothing is editable from the application, and it states plainly that nothing is
collected, transmitted or retained.

Video import claimed to resample to 24 kHz but only relabelled the header, so a
48 kHz source — every phone and screen recording — came out twice as long and an
octave too low. FFmpeg now performs the conversion; verified at 48, 44.1, 24 and
16 kHz.

"Le runtime qwen-tts manque" could not be escaped because availability was
probed with find_spec, which reports a module a packaged build cannot import, so
the existing fallback never triggered. The probe performs the real import once
and caches it, and engines missing from the build are shown disabled.

Shortcut capture needed a second click: Qt allows one keyboard grab and the
previous row never released it. Generated audio is no longer added to the
favorites unasked. History entries are named by what they say, show and filter
by voice, and can be deleted with their audio.

Voice fidelity follows measurements made on the VoiceClone project across 72
takes with the same engines: a median-pitch guard regenerates a take more than 6
semitones from its reference, and cloning defaults to 0.7x. Pocket TTS exposes
no tempo control and the existing speed setting was silently discarded, so the
correction is applied after generation with a pitch-preserving overlap-add — a
plain resample would have slowed the voice by dropping it an octave.

Also: the drive scan no longer probes a mapped network share before checking its
type, which blocked for the SMB timeout, and a test that started a real
generation no longer outlives itself and crashes the interpreter.

+++ src/soundmaster/core/fast_audio.py
@@ -43,7 +43,7 @@ def load_audio_pcm(path: Path | str, target_sr: int = 44100) -> tuple[np.ndarray
         data = data[:, :2]
 
     if sr != target_sr:
-        num_samples = int(round(len(data) * target_sr / sr))
+        num_samples = round(len(data) * target_sr / sr)
         indices = np.linspace(0, len(data) - 1, num_samples)
         data = np.array(
             [np.interp(indices, np.arange(len(data)), data[:, c]) for c in range(2)]
@@ -132,7 +132,7 @@ class ContinuousAudioOutput:
             )
             self._stream.start()
             self._retry_pending = False
-        except Exception as err:
+        except Exception as err:  # noqa: BLE001 - audio backend boundary.
             logger.warning(f"ContinuousAudioOutput stream start error: {err}")
             self._stream = None
             self._start_retry_loop()
@@ -180,7 +180,7 @@ class ContinuousAudioOutput:
             for i, dev in enumerate(sd.query_devices()):
                 if dev["max_output_channels"] > 0 and name_str in dev["name"].lower():
                     return i
-        except Exception:
+        except Exception:  # noqa: BLE001, S110 - probing devices must never raise.
             pass
         return None
 
@@ -196,7 +196,7 @@ class ContinuousAudioOutput:
             try:
                 stream_to_close.stop()
                 stream_to_close.close()
-            except Exception:
+            except Exception:  # noqa: BLE001, S110 - probing devices must never raise.
                 pass
 
         with self._lock:
@@ -256,7 +256,7 @@ class ContinuousAudioOutput:
             try:
                 stream_to_close.stop()
                 stream_to_close.close()
-            except Exception:
+            except Exception:  # noqa: BLE001, S110 - probing devices must never raise.
                 pass
         if self._retry_thread is not None:
             self._retry_thread.join(timeout=2.0)
@@ -279,7 +279,7 @@ class FastAudioEngine:
             pcm, _ = load_audio_pcm(key)
             self._pcm_cache[key] = pcm
             return True
-        except Exception as err:
+        except Exception as err:  # noqa: BLE001 - audio backend boundary.
             logger.warning(f"Preload audio failed for {path}: {err}")
             return False
 
@@ -306,7 +306,7 @@ class FastAudioEngine:
             try:
                 pcm, _ = load_audio_pcm(key)
                 self._pcm_cache[key] = pcm
-            except Exception as err:
+            except Exception as err:  # noqa: BLE001 - audio backend boundary.
                 logger.error(f"Play failed for {path}: {err}")
                 return False
